% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{BioinformaticsMethodInfo}
\alias{BioinformaticsMethodInfo}
\title{BioinformaticsMethodInfo}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
The targeted amplicon bioinformatics methods used to generate the microhaplotype data in this PMO.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{methods}: Methodology used to generate the microhaplotype data stored in this PMO, e.g. demultiplexing method, denosing method, or a pipeline method that ties all th steps together.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{methods}}{Methodology used to generate the microhaplotype data stored in this PMO, e.g. demultiplexing method, denosing method, or a pipeline method that ties all th steps together.}

    \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-BioinformaticsMethodInfo-initialize}{\code{BioinformaticsMethodInfo$new()}}
    \item \href{#method-BioinformaticsMethodInfo-validate}{\code{BioinformaticsMethodInfo$validate()}}
    \item \href{#method-BioinformaticsMethodInfo-to_list}{\code{BioinformaticsMethodInfo$to_list()}}
    \item \href{#method-BioinformaticsMethodInfo-to_json_list}{\code{BioinformaticsMethodInfo$to_json_list()}}
    \item \href{#method-BioinformaticsMethodInfo-to_json}{\code{BioinformaticsMethodInfo$to_json()}}
    \item \href{#method-BioinformaticsMethodInfo-clone}{\code{BioinformaticsMethodInfo$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioinformaticsMethodInfo-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-BioinformaticsMethodInfo-initialize}{}}}
\subsection{\code{BioinformaticsMethodInfo$new()}}{
  Create a new instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioinformaticsMethodInfo$new(methods = list(), extras = list())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{methods}}{Methodology used to generate the microhaplotype data stored in this PMO, e.g. demultiplexing method, denosing method, or a pipeline method that ties all th steps together.}
      \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioinformaticsMethodInfo-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BioinformaticsMethodInfo-validate}{}}}
\subsection{\code{BioinformaticsMethodInfo$validate()}}{
  Validate the current instance against schema-derived constraints.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioinformaticsMethodInfo$validate()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioinformaticsMethodInfo-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-BioinformaticsMethodInfo-to_list}{}}}
\subsection{\code{BioinformaticsMethodInfo$to_list()}}{
  Convert the object to a plain R list using in-memory values.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioinformaticsMethodInfo$to_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioinformaticsMethodInfo-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-BioinformaticsMethodInfo-to_json_list}{}}}
\subsection{\code{BioinformaticsMethodInfo$to_json_list()}}{
  Convert the object to a JSON-ready R list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioinformaticsMethodInfo$to_json_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioinformaticsMethodInfo-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-BioinformaticsMethodInfo-to_json}{}}}
\subsection{\code{BioinformaticsMethodInfo$to_json()}}{
  Convert the object to a JSON string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioinformaticsMethodInfo$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pretty}}{Logical; pretty-print the JSON.}
      \item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
      \item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioinformaticsMethodInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BioinformaticsMethodInfo-clone}{}}}
\subsection{\code{BioinformaticsMethodInfo$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioinformaticsMethodInfo$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
