% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_builder_merge.R
\name{.merge_replace_key}
\alias{.merge_replace_key}
\title{Pop name_key from each entry and set id_key from a (1-based) lookup}
\usage{
.merge_replace_key(target_list, lookup, name_key, id_key)
}
\description{
Pop name_key from each entry and set id_key from a (1-based) lookup
}
\keyword{internal}
