% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_builder_metatable.R
\name{.meta_build_records}
\alias{.meta_build_records}
\title{Build name->field records from a metadata table}
\usage{
.meta_build_records(wdf, fields, required_fields)
}
\description{
Build name->field records from a metadata table
}
\keyword{internal}
