% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_checker.R
\name{.pmo_resolve_schema_text}
\alias{.pmo_resolve_schema_text}
\title{Resolve a schema argument to schema text for jsonvalidate}
\usage{
.pmo_resolve_schema_text(schema = NULL, version = NULL)
}
\description{
Resolve a schema argument to schema text for jsonvalidate
}
\keyword{internal}
