% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_checker.R
\name{.pmo_schema_path}
\alias{.pmo_schema_path}
\title{Path to a bundled PMO schema file}
\usage{
.pmo_schema_path(version = pmo_schema_version())
}
\description{
Path to a bundled PMO schema file
}
\keyword{internal}
