% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{open_text_connection}
\alias{open_text_connection}
\title{Open a text connection, optionally compressed}
\usage{
open_text_connection(path, mode = c("rt", "wt"))
}
\arguments{
\item{path}{File path.}

\item{mode}{Either \code{"rt"} or \code{"wt"}.}
}
\value{
A connection object.
}
\description{
Open a text connection, optionally compressed
}
\keyword{internal}
