% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{PanelInfo}
\alias{PanelInfo}
\title{PanelInfo}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
Information on a panel of targeted amplicon primer pairs.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{panel_name}: A name for the panel.
\item \code{reactions}: A list of 1 or more reactions that this panel contains, each reactions list the targets that were amplified in that reaction, e.g. pool1, pool2.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{panel_name}}{A name for the panel.}

    \item{\code{reactions}}{A list of 1 or more reactions that this panel contains, each reactions list the targets that were amplified in that reaction, e.g. pool1, pool2.}

    \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PanelInfo-initialize}{\code{PanelInfo$new()}}
    \item \href{#method-PanelInfo-validate}{\code{PanelInfo$validate()}}
    \item \href{#method-PanelInfo-to_list}{\code{PanelInfo$to_list()}}
    \item \href{#method-PanelInfo-to_json_list}{\code{PanelInfo$to_json_list()}}
    \item \href{#method-PanelInfo-to_json}{\code{PanelInfo$to_json()}}
    \item \href{#method-PanelInfo-clone}{\code{PanelInfo$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PanelInfo-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PanelInfo-initialize}{}}}
\subsection{\code{PanelInfo$new()}}{
  Create a new instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PanelInfo$new(panel_name = NA_character_, reactions = list(), extras = list())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{panel_name}}{A name for the panel.}
      \item{\code{reactions}}{A list of 1 or more reactions that this panel contains, each reactions list the targets that were amplified in that reaction, e.g. pool1, pool2.}
      \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PanelInfo-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PanelInfo-validate}{}}}
\subsection{\code{PanelInfo$validate()}}{
  Validate the current instance against schema-derived constraints.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PanelInfo$validate()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PanelInfo-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-PanelInfo-to_list}{}}}
\subsection{\code{PanelInfo$to_list()}}{
  Convert the object to a plain R list using in-memory values.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PanelInfo$to_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PanelInfo-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-PanelInfo-to_json_list}{}}}
\subsection{\code{PanelInfo$to_json_list()}}{
  Convert the object to a JSON-ready R list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PanelInfo$to_json_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PanelInfo-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-PanelInfo-to_json}{}}}
\subsection{\code{PanelInfo$to_json()}}{
  Convert the object to a JSON string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PanelInfo$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pretty}}{Logical; pretty-print the JSON.}
      \item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
      \item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PanelInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PanelInfo-clone}{}}}
\subsection{\code{PanelInfo$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PanelInfo$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
