% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{ParasiteDensity}
\alias{ParasiteDensity}
\title{ParasiteDensity}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
ParasiteDensity

ParasiteDensity
}
\details{
Method and value of determined parasite density.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{date_measured}: The date the qpcr was performed, can be YYYY, YYYY-MM, or YYYY-MM-DD.
\item \code{density_method_comments}: Additional comments about how the density was performed.
\item \code{parasite_density}: The density in microliters.
\item \code{parasite_density_method}: The method of how this density was obtained.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{date_measured}}{The date the qpcr was performed, can be YYYY, YYYY-MM, or YYYY-MM-DD.}

\item{\code{density_method_comments}}{Additional comments about how the density was performed.}

\item{\code{parasite_density}}{The density in microliters.}

\item{\code{parasite_density_method}}{The method of how this density was obtained.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParasiteDensity-new}{\code{ParasiteDensity$new()}}
\item \href{#method-ParasiteDensity-validate}{\code{ParasiteDensity$validate()}}
\item \href{#method-ParasiteDensity-to_list}{\code{ParasiteDensity$to_list()}}
\item \href{#method-ParasiteDensity-to_json_list}{\code{ParasiteDensity$to_json_list()}}
\item \href{#method-ParasiteDensity-to_json}{\code{ParasiteDensity$to_json()}}
\item \href{#method-ParasiteDensity-clone}{\code{ParasiteDensity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParasiteDensity-new"></a>}}
\if{latex}{\out{\hypertarget{method-ParasiteDensity-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParasiteDensity$new(
  date_measured = NULL,
  density_method_comments = NULL,
  parasite_density = NA_real_,
  parasite_density_method = NA_character_,
  extras = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date_measured}}{The date the qpcr was performed, can be YYYY, YYYY-MM, or YYYY-MM-DD.}

\item{\code{density_method_comments}}{Additional comments about how the density was performed.}

\item{\code{parasite_density}}{The density in microliters.}

\item{\code{parasite_density_method}}{The method of how this density was obtained.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParasiteDensity-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ParasiteDensity-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParasiteDensity$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParasiteDensity-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ParasiteDensity-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParasiteDensity$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParasiteDensity-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-ParasiteDensity-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParasiteDensity$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParasiteDensity-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-ParasiteDensity-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParasiteDensity$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParasiteDensity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ParasiteDensity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParasiteDensity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
