% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_processor.R
\name{pmo_count_targets_per_panel}
\alias{pmo_count_targets_per_panel}
\title{Count the number of unique targets in each panel}
\usage{
pmo_count_targets_per_panel(pmo)
}
\arguments{
\item{pmo}{A \code{PortableMicrohaplotypeObject} or parsed PMO list.}
}
\value{
A tibble with columns \code{panel_name}, \code{panel_target_count}.
}
\description{
Count the number of unique targets in each panel
}
\examples{
p <- read_pmo(
  system.file("extdata", "example_full_pmo.json.gz", package = "pmotoolsr"))
pmo_count_targets_per_panel(p)
}
