% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_exporter.R
\name{pmo_export_bioinformatics_run_info_meta_table}
\alias{pmo_export_bioinformatics_run_info_meta_table}
\title{Export bioinformatics run info metadata}
\usage{
pmo_export_bioinformatics_run_info_meta_table(pmo, separator = ",")
}
\arguments{
\item{pmo}{A \code{PortableMicrohaplotypeObject} or parsed PMO list.}

\item{separator}{Separator for list-valued fields.}
}
\value{
A tibble with a 1-based \code{run_id} column.
}
\description{
Export bioinformatics run info metadata
}
\examples{
p <- read_pmo(
  system.file("extdata", "example_full_pmo.json.gz", package = "pmotoolsr"))
pmo_export_bioinformatics_run_info_meta_table(p)
}
