% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_exporter.R
\name{pmo_export_library_sample_meta_table}
\alias{pmo_export_library_sample_meta_table}
\title{Export library sample metadata}
\usage{
pmo_export_library_sample_meta_table(pmo, separator = ",")
}
\arguments{
\item{pmo}{A \code{PortableMicrohaplotypeObject} or parsed PMO list.}

\item{separator}{Separator for list-valued fields.}
}
\value{
A tibble (\code{sequencing_info_id}/\code{specimen_id}/\code{panel_id} resolved to
their names).
}
\description{
Export library sample metadata
}
\examples{
p <- read_pmo(
  system.file("extdata", "example_full_pmo.json.gz", package = "pmotoolsr"))
head(pmo_export_library_sample_meta_table(p))
}
