% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_exporter.R
\name{pmo_export_pmo_header_table}
\alias{pmo_export_pmo_header_table}
\title{Export PMO header metadata}
\usage{
pmo_export_pmo_header_table(pmo, separator = ",")
}
\arguments{
\item{pmo}{A \code{PortableMicrohaplotypeObject} or parsed PMO list.}

\item{separator}{Separator for list-valued fields.}
}
\value{
A one-row tibble.
}
\description{
Export PMO header metadata
}
\examples{
p <- read_pmo(
  system.file("extdata", "example_full_pmo.json.gz", package = "pmotoolsr"))
pmo_export_pmo_header_table(p)
}
