% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_processor.R
\name{pmo_filter_by_specimen_ids}
\alias{pmo_filter_by_specimen_ids}
\title{Filter a PMO down to selected specimens}
\usage{
pmo_filter_by_specimen_ids(pmo, specimen_ids)
}
\arguments{
\item{pmo}{A \code{PortableMicrohaplotypeObject} or parsed PMO list.}

\item{specimen_ids}{Integer vector of 1-based specimen ids.}
}
\value{
A PMO list (write with \code{\link[=write_pmo_raw]{write_pmo_raw()}}).
}
\description{
Filter a PMO down to selected specimens
}
\examples{
p <- read_pmo(
  system.file("extdata", "example_full_pmo.json.gz", package = "pmotoolsr"))
sub <- pmo_filter_by_specimen_ids(p, 1L)
}
