% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_processor.R
\name{pmo_filter_by_specimen_names}
\alias{pmo_filter_by_specimen_names}
\title{Filter a PMO down to selected specimens by name}
\usage{
pmo_filter_by_specimen_names(pmo, specimen_names)
}
\arguments{
\item{pmo}{A \code{PortableMicrohaplotypeObject} or parsed PMO list.}

\item{specimen_names}{Character vector of specimen names.}
}
\value{
A PMO list (write with \code{\link[=write_pmo_raw]{write_pmo_raw()}}).
}
\description{
Filter a PMO down to selected specimens by name
}
\examples{
p <- read_pmo(
  system.file("extdata", "example_full_pmo.json.gz", package = "pmotoolsr"))
nm <- pmo_get_specimen_names(p)[1]
sub <- pmo_filter_by_specimen_names(p, nm)
}
