% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_processor.R
\name{pmo_get_specimen_names}
\alias{pmo_get_specimen_names}
\alias{pmo_get_sorted_specimen_names}
\alias{pmo_get_library_sample_names}
\alias{pmo_get_sorted_library_sample_names}
\alias{pmo_get_target_names}
\alias{pmo_get_sorted_target_names}
\alias{pmo_get_panel_names}
\alias{pmo_get_sorted_panel_names}
\alias{pmo_get_bioinformatics_run_names}
\alias{pmo_get_sorted_bioinformatics_run_names}
\title{Get entity names from a PMO}
\usage{
pmo_get_specimen_names(pmo)

pmo_get_sorted_specimen_names(pmo)

pmo_get_library_sample_names(pmo)

pmo_get_sorted_library_sample_names(pmo)

pmo_get_target_names(pmo)

pmo_get_sorted_target_names(pmo)

pmo_get_panel_names(pmo)

pmo_get_sorted_panel_names(pmo)

pmo_get_bioinformatics_run_names(pmo)

pmo_get_sorted_bioinformatics_run_names(pmo)
}
\arguments{
\item{pmo}{A \code{PortableMicrohaplotypeObject} or parsed PMO list.}
}
\value{
A character vector of names.
}
\description{
Return the names of PMO entities in the order they appear. Each function has
a \verb{pmo_get_sorted_*} companion returning the names sorted alphabetically.
}
\examples{
pmo <- read_pmo(
  system.file("extdata", "example_pmo.json.gz", package = "pmotoolsr"))
head(pmo_get_specimen_names(pmo))
pmo_get_sorted_target_names(pmo)[1:3]
}
