% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_processor.R
\name{pmo_has_section}
\alias{pmo_has_section}
\title{Test whether an optional PMO section is present and non-empty}
\usage{
pmo_has_section(pmo, section)
}
\arguments{
\item{pmo}{A \code{PortableMicrohaplotypeObject} or parsed PMO list.}

\item{section}{The name of a top-level PMO section.}
}
\value{
\code{TRUE} if the section is present and contains at least one element.
}
\description{
Test whether an optional PMO section is present and non-empty
}
\examples{
p <- read_pmo(
  system.file("extdata", "example_full_pmo.json.gz", package = "pmotoolsr"))
pmo_has_section(p, "sequencing_info")
}
