% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_processor.R
\name{pmo_index_of_specimen_names}
\alias{pmo_index_of_specimen_names}
\alias{pmo_index_of_library_sample_names}
\alias{pmo_index_of_target_names}
\alias{pmo_index_of_panel_names}
\alias{pmo_index_of_bioinformatics_run_names}
\alias{pmo_index_of_target_in_representative_microhaplotypes}
\title{Resolve entity names to their 1-based PMO indices}
\usage{
pmo_index_of_specimen_names(pmo, specimen_names)

pmo_index_of_library_sample_names(pmo, library_sample_names)

pmo_index_of_target_names(pmo, target_names)

pmo_index_of_panel_names(pmo, panel_names)

pmo_index_of_bioinformatics_run_names(pmo, bioinformatics_run_names)

pmo_index_of_target_in_representative_microhaplotypes(pmo, target_names)
}
\arguments{
\item{pmo}{A \code{PortableMicrohaplotypeObject} or parsed PMO list.}

\item{specimen_names, library_sample_names, target_names, panel_names, bioinformatics_run_names}{A character vector of names to resolve.}
}
\value{
An integer vector of 1-based indices.
}
\description{
Convert a vector of names into the corresponding 1-based indices, returned in
the same order as the input. Unknown names raise an error.
}
\examples{
p <- read_pmo(
  system.file("extdata", "example_full_pmo.json.gz", package = "pmotoolsr"))
pmo_index_of_specimen_names(p, pmo_get_specimen_names(p)[1:2])
}
