% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_checker.R
\name{pmo_schema_version}
\alias{pmo_schema_version}
\title{Default PMO schema version targeted by this package}
\usage{
pmo_schema_version()
}
\value{
A version string.
}
\description{
The R6 classes in this package were generated from this schema version.
}
\examples{
pmo_schema_version()
}
