% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{PmoGenerationMethod}
\alias{PmoGenerationMethod}
\title{PmoGenerationMethod}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
PmoGenerationMethod

PmoGenerationMethod
}
\details{
Information about how a PMO was generated.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{program_name}: The name of the program.
\item \code{program_version}: The version of program, should be in the format of v[MAJOR].[MINOR].[PATCH].
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{program_name}}{The name of the program.}

\item{\code{program_version}}{The version of program, should be in the format of v[MAJOR].[MINOR].[PATCH].}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PmoGenerationMethod-new}{\code{PmoGenerationMethod$new()}}
\item \href{#method-PmoGenerationMethod-validate}{\code{PmoGenerationMethod$validate()}}
\item \href{#method-PmoGenerationMethod-to_list}{\code{PmoGenerationMethod$to_list()}}
\item \href{#method-PmoGenerationMethod-to_json_list}{\code{PmoGenerationMethod$to_json_list()}}
\item \href{#method-PmoGenerationMethod-to_json}{\code{PmoGenerationMethod$to_json()}}
\item \href{#method-PmoGenerationMethod-clone}{\code{PmoGenerationMethod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PmoGenerationMethod-new"></a>}}
\if{latex}{\out{\hypertarget{method-PmoGenerationMethod-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PmoGenerationMethod$new(
  program_name = NA_character_,
  program_version = NA_character_,
  extras = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{program_name}}{The name of the program.}

\item{\code{program_version}}{The version of program, should be in the format of v[MAJOR].[MINOR].[PATCH].}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PmoGenerationMethod-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PmoGenerationMethod-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PmoGenerationMethod$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PmoGenerationMethod-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-PmoGenerationMethod-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PmoGenerationMethod$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PmoGenerationMethod-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-PmoGenerationMethod-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PmoGenerationMethod$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PmoGenerationMethod-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-PmoGenerationMethod-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PmoGenerationMethod$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PmoGenerationMethod-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PmoGenerationMethod-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PmoGenerationMethod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
