% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{PmoHeader}
\alias{PmoHeader}
\title{PmoHeader}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
PmoHeader

PmoHeader
}
\details{
Information on the PMO file itself.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{creation_date}: The date of when the PMO file was created or modified, should be YYYY-MM-DD.
\item \code{generation_method}: The generation method to create this PMO.
\item \code{pmo_version}: The version of the PMO file, should be in the format of v[MAJOR].[MINOR].[PATCH].
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{creation_date}}{The date of when the PMO file was created or modified, should be YYYY-MM-DD.}

\item{\code{generation_method}}{The generation method to create this PMO.}

\item{\code{pmo_version}}{The version of the PMO file, should be in the format of v[MAJOR].[MINOR].[PATCH].}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PmoHeader-new}{\code{PmoHeader$new()}}
\item \href{#method-PmoHeader-validate}{\code{PmoHeader$validate()}}
\item \href{#method-PmoHeader-to_list}{\code{PmoHeader$to_list()}}
\item \href{#method-PmoHeader-to_json_list}{\code{PmoHeader$to_json_list()}}
\item \href{#method-PmoHeader-to_json}{\code{PmoHeader$to_json()}}
\item \href{#method-PmoHeader-clone}{\code{PmoHeader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PmoHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-PmoHeader-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PmoHeader$new(
  creation_date = NULL,
  generation_method = NULL,
  pmo_version = NA_character_,
  extras = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{creation_date}}{The date of when the PMO file was created or modified, should be YYYY-MM-DD.}

\item{\code{generation_method}}{The generation method to create this PMO.}

\item{\code{pmo_version}}{The version of the PMO file, should be in the format of v[MAJOR].[MINOR].[PATCH].}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PmoHeader-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PmoHeader-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PmoHeader$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PmoHeader-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-PmoHeader-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PmoHeader$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PmoHeader-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-PmoHeader-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PmoHeader$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PmoHeader-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-PmoHeader-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PmoHeader$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PmoHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PmoHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PmoHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
