% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmotoolsr-package.R
\docType{package}
\name{pmotoolsr-package}
\alias{pmotoolsr}
\alias{pmotoolsr-package}
\title{pmotoolsr: Toolkit for working with Portable Microhaplotype Objects (PMOs)}
\description{
Read, write, validate, process, build, and combine Portable Microhaplotype
Object (PMO) files in R. See the README and the function reference for
details, and note the 0-based (on disk) to 1-based (in R) index convention.
}
\author{
\strong{Maintainer}: Nicholas Hathaway \email{nicholas.hathaway@ucsf.edu} (\href{https://orcid.org/0000-0001-9639-2894}{ORCID})

Authors:
\itemize{
  \item Nicholas Hathaway \email{nicholas.hathaway@ucsf.edu} (\href{https://orcid.org/0000-0001-9639-2894}{ORCID})
}

}
\keyword{internal}
