% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{PrimerInfo}
\alias{PrimerInfo}
\title{PrimerInfo}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
Information on a primer sequence.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{location}: What the intended genomic location of the primer is.
\item \code{seq}: The sequence.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{location}}{What the intended genomic location of the primer is.}

    \item{\code{seq}}{The sequence.}

    \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PrimerInfo-initialize}{\code{PrimerInfo$new()}}
    \item \href{#method-PrimerInfo-validate}{\code{PrimerInfo$validate()}}
    \item \href{#method-PrimerInfo-to_list}{\code{PrimerInfo$to_list()}}
    \item \href{#method-PrimerInfo-to_json_list}{\code{PrimerInfo$to_json_list()}}
    \item \href{#method-PrimerInfo-to_json}{\code{PrimerInfo$to_json()}}
    \item \href{#method-PrimerInfo-clone}{\code{PrimerInfo$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimerInfo-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PrimerInfo-initialize}{}}}
\subsection{\code{PrimerInfo$new()}}{
  Create a new instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PrimerInfo$new(location = NULL, seq = NA_character_, extras = list())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{location}}{What the intended genomic location of the primer is.}
      \item{\code{seq}}{The sequence.}
      \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimerInfo-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PrimerInfo-validate}{}}}
\subsection{\code{PrimerInfo$validate()}}{
  Validate the current instance against schema-derived constraints.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PrimerInfo$validate()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimerInfo-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-PrimerInfo-to_list}{}}}
\subsection{\code{PrimerInfo$to_list()}}{
  Convert the object to a plain R list using in-memory values.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PrimerInfo$to_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimerInfo-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-PrimerInfo-to_json_list}{}}}
\subsection{\code{PrimerInfo$to_json_list()}}{
  Convert the object to a JSON-ready R list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PrimerInfo$to_json_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimerInfo-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-PrimerInfo-to_json}{}}}
\subsection{\code{PrimerInfo$to_json()}}{
  Convert the object to a JSON string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PrimerInfo$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pretty}}{Logical; pretty-print the JSON.}
      \item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
      \item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PrimerInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PrimerInfo-clone}{}}}
\subsection{\code{PrimerInfo$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PrimerInfo$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
