% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{ReactionInfo}
\alias{ReactionInfo}
\title{ReactionInfo}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
ReactionInfo

ReactionInfo
}
\details{
Information on a panel of targeted amplicon primer pairs.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{panel_targets}: A list of the target indexes in the target_info list.
\item \code{reaction_name}: A name for this reaction.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{panel_targets}}{A list of the target indexes in the target_info list.}

\item{\code{reaction_name}}{A name for this reaction.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReactionInfo-new}{\code{ReactionInfo$new()}}
\item \href{#method-ReactionInfo-validate}{\code{ReactionInfo$validate()}}
\item \href{#method-ReactionInfo-to_list}{\code{ReactionInfo$to_list()}}
\item \href{#method-ReactionInfo-to_json_list}{\code{ReactionInfo$to_json_list()}}
\item \href{#method-ReactionInfo-to_json}{\code{ReactionInfo$to_json()}}
\item \href{#method-ReactionInfo-clone}{\code{ReactionInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactionInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReactionInfo-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionInfo$new(
  panel_targets = numeric(),
  reaction_name = NA_character_,
  extras = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{panel_targets}}{A list of the target indexes in the target_info list.}

\item{\code{reaction_name}}{A name for this reaction.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactionInfo-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ReactionInfo-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionInfo$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactionInfo-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ReactionInfo-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionInfo$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactionInfo-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-ReactionInfo-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionInfo$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactionInfo-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-ReactionInfo-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionInfo$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactionInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReactionInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactionInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
