% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{StageReadCounts}
\alias{StageReadCounts}
\title{StageReadCounts}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
StageReadCounts

StageReadCounts
}
\details{
Information on the reads counts at several stages.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{reads}: The read counts for this stage.
\item \code{stage}: The stage of the pipeline, e.g. demultiplexed, denoised, etc.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{reads}}{The read counts for this stage.}

\item{\code{stage}}{The stage of the pipeline, e.g. demultiplexed, denoised, etc.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StageReadCounts-new}{\code{StageReadCounts$new()}}
\item \href{#method-StageReadCounts-validate}{\code{StageReadCounts$validate()}}
\item \href{#method-StageReadCounts-to_list}{\code{StageReadCounts$to_list()}}
\item \href{#method-StageReadCounts-to_json_list}{\code{StageReadCounts$to_json_list()}}
\item \href{#method-StageReadCounts-to_json}{\code{StageReadCounts$to_json()}}
\item \href{#method-StageReadCounts-clone}{\code{StageReadCounts$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StageReadCounts-new"></a>}}
\if{latex}{\out{\hypertarget{method-StageReadCounts-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StageReadCounts$new(reads = NA_real_, stage = NA_character_, extras = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reads}}{The read counts for this stage.}

\item{\code{stage}}{The stage of the pipeline, e.g. demultiplexed, denoised, etc.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StageReadCounts-validate"></a>}}
\if{latex}{\out{\hypertarget{method-StageReadCounts-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StageReadCounts$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StageReadCounts-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-StageReadCounts-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StageReadCounts$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StageReadCounts-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-StageReadCounts-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StageReadCounts$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StageReadCounts-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-StageReadCounts-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StageReadCounts$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StageReadCounts-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StageReadCounts-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StageReadCounts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
