% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{BioinformaticsRunInfo}
\alias{BioinformaticsRunInfo}
\title{BioinformaticsRunInfo}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
Information about the pipeline run that generated microhaplotype_detected and reads_by_stage.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{bioinformatics_methods_id}: The index into the bioinformatics_methods_info list.
\item \code{bioinformatics_run_name}: A name to for this run, needs to be unique to each run.
\item \code{run_date}: The date when the run was done, should be YYYY-MM-DD.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{bioinformatics_methods_id}}{The index into the bioinformatics_methods_info list.}

    \item{\code{bioinformatics_run_name}}{A name to for this run, needs to be unique to each run.}

    \item{\code{run_date}}{The date when the run was done, should be YYYY-MM-DD.}

    \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-BioinformaticsRunInfo-initialize}{\code{BioinformaticsRunInfo$new()}}
    \item \href{#method-BioinformaticsRunInfo-validate}{\code{BioinformaticsRunInfo$validate()}}
    \item \href{#method-BioinformaticsRunInfo-to_list}{\code{BioinformaticsRunInfo$to_list()}}
    \item \href{#method-BioinformaticsRunInfo-to_json_list}{\code{BioinformaticsRunInfo$to_json_list()}}
    \item \href{#method-BioinformaticsRunInfo-to_json}{\code{BioinformaticsRunInfo$to_json()}}
    \item \href{#method-BioinformaticsRunInfo-clone}{\code{BioinformaticsRunInfo$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioinformaticsRunInfo-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-BioinformaticsRunInfo-initialize}{}}}
\subsection{\code{BioinformaticsRunInfo$new()}}{
  Create a new instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioinformaticsRunInfo$new(
  bioinformatics_methods_id = NA_real_,
  bioinformatics_run_name = NA_character_,
  run_date = NULL,
  extras = list()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{bioinformatics_methods_id}}{The index into the bioinformatics_methods_info list.}
      \item{\code{bioinformatics_run_name}}{A name to for this run, needs to be unique to each run.}
      \item{\code{run_date}}{The date when the run was done, should be YYYY-MM-DD.}
      \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioinformaticsRunInfo-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BioinformaticsRunInfo-validate}{}}}
\subsection{\code{BioinformaticsRunInfo$validate()}}{
  Validate the current instance against schema-derived constraints.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioinformaticsRunInfo$validate()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioinformaticsRunInfo-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-BioinformaticsRunInfo-to_list}{}}}
\subsection{\code{BioinformaticsRunInfo$to_list()}}{
  Convert the object to a plain R list using in-memory values.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioinformaticsRunInfo$to_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioinformaticsRunInfo-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-BioinformaticsRunInfo-to_json_list}{}}}
\subsection{\code{BioinformaticsRunInfo$to_json_list()}}{
  Convert the object to a JSON-ready R list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioinformaticsRunInfo$to_json_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioinformaticsRunInfo-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-BioinformaticsRunInfo-to_json}{}}}
\subsection{\code{BioinformaticsRunInfo$to_json()}}{
  Convert the object to a JSON string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioinformaticsRunInfo$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pretty}}{Logical; pretty-print the JSON.}
      \item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
      \item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioinformaticsRunInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BioinformaticsRunInfo-clone}{}}}
\subsection{\code{BioinformaticsRunInfo$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioinformaticsRunInfo$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
