% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{BioMethod}
\alias{BioMethod}
\title{BioMethod}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
Bioinformatics methodology description with info on program, version, and arguments different from the default.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{additional_argument}: Any additional arguments that differ from the default arguments.
\item \code{program}: Name of the program used for this portion of the pipeline.
\item \code{program_description}: A short description of what this method does.
\item \code{program_url}: A url pointing to code base of a program, e.g. a github link.
\item \code{program_version}: The version of program, should be in the format of v[MAJOR].[MINOR].[PATCH].
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{additional_argument}}{Any additional arguments that differ from the default arguments.}

    \item{\code{program}}{Name of the program used for this portion of the pipeline.}

    \item{\code{program_description}}{A short description of what this method does.}

    \item{\code{program_url}}{A url pointing to code base of a program, e.g. a github link.}

    \item{\code{program_version}}{The version of program, should be in the format of v[MAJOR].[MINOR].[PATCH].}

    \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-BioMethod-initialize}{\code{BioMethod$new()}}
    \item \href{#method-BioMethod-validate}{\code{BioMethod$validate()}}
    \item \href{#method-BioMethod-to_list}{\code{BioMethod$to_list()}}
    \item \href{#method-BioMethod-to_json_list}{\code{BioMethod$to_json_list()}}
    \item \href{#method-BioMethod-to_json}{\code{BioMethod$to_json()}}
    \item \href{#method-BioMethod-clone}{\code{BioMethod$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioMethod-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-BioMethod-initialize}{}}}
\subsection{\code{BioMethod$new()}}{
  Create a new instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioMethod$new(
  additional_argument = NULL,
  program = NA_character_,
  program_description = NULL,
  program_url = NULL,
  program_version = NA_character_,
  extras = list()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{additional_argument}}{Any additional arguments that differ from the default arguments.}
      \item{\code{program}}{Name of the program used for this portion of the pipeline.}
      \item{\code{program_description}}{A short description of what this method does.}
      \item{\code{program_url}}{A url pointing to code base of a program, e.g. a github link.}
      \item{\code{program_version}}{The version of program, should be in the format of v[MAJOR].[MINOR].[PATCH].}
      \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioMethod-validate"></a>}}
\if{latex}{\out{\hypertarget{method-BioMethod-validate}{}}}
\subsection{\code{BioMethod$validate()}}{
  Validate the current instance against schema-derived constraints.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioMethod$validate()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioMethod-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-BioMethod-to_list}{}}}
\subsection{\code{BioMethod$to_list()}}{
  Convert the object to a plain R list using in-memory values.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioMethod$to_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioMethod-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-BioMethod-to_json_list}{}}}
\subsection{\code{BioMethod$to_json_list()}}{
  Convert the object to a JSON-ready R list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioMethod$to_json_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioMethod-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-BioMethod-to_json}{}}}
\subsection{\code{BioMethod$to_json()}}{
  Convert the object to a JSON string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioMethod$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pretty}}{Logical; pretty-print the JSON.}
      \item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
      \item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BioMethod-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BioMethod-clone}{}}}
\subsection{\code{BioMethod$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{BioMethod$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
