% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{DetectedMicrohaplotypesForSample}
\alias{DetectedMicrohaplotypesForSample}
\title{DetectedMicrohaplotypesForSample}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
Microhaplotypes detected for a sample for all targets.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{library_sample_id}: The index into the library_sample_info list.
\item \code{target_results}: A list of the microhaplotypes detected for a list of targets.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{library_sample_id}}{The index into the library_sample_info list.}

    \item{\code{target_results}}{A list of the microhaplotypes detected for a list of targets.}

    \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DetectedMicrohaplotypesForSample-initialize}{\code{DetectedMicrohaplotypesForSample$new()}}
    \item \href{#method-DetectedMicrohaplotypesForSample-validate}{\code{DetectedMicrohaplotypesForSample$validate()}}
    \item \href{#method-DetectedMicrohaplotypesForSample-to_list}{\code{DetectedMicrohaplotypesForSample$to_list()}}
    \item \href{#method-DetectedMicrohaplotypesForSample-to_json_list}{\code{DetectedMicrohaplotypesForSample$to_json_list()}}
    \item \href{#method-DetectedMicrohaplotypesForSample-to_json}{\code{DetectedMicrohaplotypesForSample$to_json()}}
    \item \href{#method-DetectedMicrohaplotypesForSample-clone}{\code{DetectedMicrohaplotypesForSample$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectedMicrohaplotypesForSample-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DetectedMicrohaplotypesForSample-initialize}{}}}
\subsection{\code{DetectedMicrohaplotypesForSample$new()}}{
  Create a new instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DetectedMicrohaplotypesForSample$new(
  library_sample_id = NA_real_,
  target_results = list(),
  extras = list()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{library_sample_id}}{The index into the library_sample_info list.}
      \item{\code{target_results}}{A list of the microhaplotypes detected for a list of targets.}
      \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectedMicrohaplotypesForSample-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DetectedMicrohaplotypesForSample-validate}{}}}
\subsection{\code{DetectedMicrohaplotypesForSample$validate()}}{
  Validate the current instance against schema-derived constraints.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DetectedMicrohaplotypesForSample$validate()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectedMicrohaplotypesForSample-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DetectedMicrohaplotypesForSample-to_list}{}}}
\subsection{\code{DetectedMicrohaplotypesForSample$to_list()}}{
  Convert the object to a plain R list using in-memory values.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DetectedMicrohaplotypesForSample$to_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectedMicrohaplotypesForSample-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-DetectedMicrohaplotypesForSample-to_json_list}{}}}
\subsection{\code{DetectedMicrohaplotypesForSample$to_json_list()}}{
  Convert the object to a JSON-ready R list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DetectedMicrohaplotypesForSample$to_json_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectedMicrohaplotypesForSample-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-DetectedMicrohaplotypesForSample-to_json}{}}}
\subsection{\code{DetectedMicrohaplotypesForSample$to_json()}}{
  Convert the object to a JSON string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DetectedMicrohaplotypesForSample$to_json(
  pretty = FALSE,
  auto_unbox = TRUE,
  ...
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pretty}}{Logical; pretty-print the JSON.}
      \item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
      \item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DetectedMicrohaplotypesForSample-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DetectedMicrohaplotypesForSample-clone}{}}}
\subsection{\code{DetectedMicrohaplotypesForSample$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DetectedMicrohaplotypesForSample$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
