% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{GenomeInfo}
\alias{GenomeInfo}
\title{GenomeInfo}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
GenomeInfo

GenomeInfo
}
\details{
Information on a genome.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{chromosomes}: A list of the chromosomes/contigs found within this genome.
\item \code{genome_version}: The genome version.
\item \code{gff_url}: A link to the where this genome's annotation file could be downloaded.
\item \code{name}: Name of the genome.
\item \code{taxon_id}: The NCBI taxonomy number, can be a list of values if it's a genome file that has been created by combining gnomes from different species.
\item \code{url}: A link to the where this genome file could be downloaded.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{chromosomes}}{A list of the chromosomes/contigs found within this genome.}

\item{\code{genome_version}}{The genome version.}

\item{\code{gff_url}}{A link to the where this genome's annotation file could be downloaded.}

\item{\code{name}}{Name of the genome.}

\item{\code{taxon_id}}{The NCBI taxonomy number, can be a list of values if it's a genome file that has been created by combining gnomes from different species.}

\item{\code{url}}{A link to the where this genome file could be downloaded.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GenomeInfo-new}{\code{GenomeInfo$new()}}
\item \href{#method-GenomeInfo-validate}{\code{GenomeInfo$validate()}}
\item \href{#method-GenomeInfo-to_list}{\code{GenomeInfo$to_list()}}
\item \href{#method-GenomeInfo-to_json_list}{\code{GenomeInfo$to_json_list()}}
\item \href{#method-GenomeInfo-to_json}{\code{GenomeInfo$to_json()}}
\item \href{#method-GenomeInfo-clone}{\code{GenomeInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomeInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-GenomeInfo-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeInfo$new(
  chromosomes = NULL,
  genome_version = NA_character_,
  gff_url = NULL,
  name = NA_character_,
  taxon_id = numeric(),
  url = NA_character_,
  extras = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chromosomes}}{A list of the chromosomes/contigs found within this genome.}

\item{\code{genome_version}}{The genome version.}

\item{\code{gff_url}}{A link to the where this genome's annotation file could be downloaded.}

\item{\code{name}}{Name of the genome.}

\item{\code{taxon_id}}{The NCBI taxonomy number, can be a list of values if it's a genome file that has been created by combining gnomes from different species.}

\item{\code{url}}{A link to the where this genome file could be downloaded.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomeInfo-validate"></a>}}
\if{latex}{\out{\hypertarget{method-GenomeInfo-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeInfo$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomeInfo-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-GenomeInfo-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeInfo$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomeInfo-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-GenomeInfo-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeInfo$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomeInfo-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-GenomeInfo-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeInfo$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomeInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GenomeInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
