% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{MarkerOfInterest}
\alias{MarkerOfInterest}
\title{MarkerOfInterest}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
MarkerOfInterest

MarkerOfInterest
}
\details{
A specific genomic location of interest, e.g. drug resistance, or other phenotypical marker.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{associations}: A list of associations with this marker, e.g. SP resistance, etc.
\item \code{marker_location}: The genomic location.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{associations}}{A list of associations with this marker, e.g. SP resistance, etc.}

\item{\code{marker_location}}{The genomic location.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MarkerOfInterest-new}{\code{MarkerOfInterest$new()}}
\item \href{#method-MarkerOfInterest-validate}{\code{MarkerOfInterest$validate()}}
\item \href{#method-MarkerOfInterest-to_list}{\code{MarkerOfInterest$to_list()}}
\item \href{#method-MarkerOfInterest-to_json_list}{\code{MarkerOfInterest$to_json_list()}}
\item \href{#method-MarkerOfInterest-to_json}{\code{MarkerOfInterest$to_json()}}
\item \href{#method-MarkerOfInterest-clone}{\code{MarkerOfInterest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkerOfInterest-new"></a>}}
\if{latex}{\out{\hypertarget{method-MarkerOfInterest-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkerOfInterest$new(
  associations = NULL,
  marker_location = NULL,
  extras = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{associations}}{A list of associations with this marker, e.g. SP resistance, etc.}

\item{\code{marker_location}}{The genomic location.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkerOfInterest-validate"></a>}}
\if{latex}{\out{\hypertarget{method-MarkerOfInterest-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkerOfInterest$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkerOfInterest-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-MarkerOfInterest-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkerOfInterest$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkerOfInterest-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-MarkerOfInterest-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkerOfInterest$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkerOfInterest-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-MarkerOfInterest-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkerOfInterest$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkerOfInterest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MarkerOfInterest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkerOfInterest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
