% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{MaskingInfo}
\alias{MaskingInfo}
\title{MaskingInfo}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
Information about a subsegment of the sequence that should be masked.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{masking_generation_description}: A description of how the masking information was generated.
\item \code{replacement_size}: The size of replacement mask.
\item \code{seq_segment_size}: The size of the masking.
\item \code{seq_start}: The start of the masking.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{masking_generation_description}}{A description of how the masking information was generated.}

    \item{\code{replacement_size}}{The size of replacement mask.}

    \item{\code{seq_segment_size}}{The size of the masking.}

    \item{\code{seq_start}}{The start of the masking.}

    \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-MaskingInfo-initialize}{\code{MaskingInfo$new()}}
    \item \href{#method-MaskingInfo-validate}{\code{MaskingInfo$validate()}}
    \item \href{#method-MaskingInfo-to_list}{\code{MaskingInfo$to_list()}}
    \item \href{#method-MaskingInfo-to_json_list}{\code{MaskingInfo$to_json_list()}}
    \item \href{#method-MaskingInfo-to_json}{\code{MaskingInfo$to_json()}}
    \item \href{#method-MaskingInfo-clone}{\code{MaskingInfo$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaskingInfo-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-MaskingInfo-initialize}{}}}
\subsection{\code{MaskingInfo$new()}}{
  Create a new instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaskingInfo$new(
  masking_generation_description = NULL,
  replacement_size = NA_real_,
  seq_segment_size = NA_real_,
  seq_start = NA_real_,
  extras = list()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{masking_generation_description}}{A description of how the masking information was generated.}
      \item{\code{replacement_size}}{The size of replacement mask.}
      \item{\code{seq_segment_size}}{The size of the masking.}
      \item{\code{seq_start}}{The start of the masking.}
      \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaskingInfo-validate"></a>}}
\if{latex}{\out{\hypertarget{method-MaskingInfo-validate}{}}}
\subsection{\code{MaskingInfo$validate()}}{
  Validate the current instance against schema-derived constraints.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaskingInfo$validate()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaskingInfo-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-MaskingInfo-to_list}{}}}
\subsection{\code{MaskingInfo$to_list()}}{
  Convert the object to a plain R list using in-memory values.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaskingInfo$to_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaskingInfo-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-MaskingInfo-to_json_list}{}}}
\subsection{\code{MaskingInfo$to_json_list()}}{
  Convert the object to a JSON-ready R list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaskingInfo$to_json_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaskingInfo-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-MaskingInfo-to_json}{}}}
\subsection{\code{MaskingInfo$to_json()}}{
  Convert the object to a JSON string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaskingInfo$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pretty}}{Logical; pretty-print the JSON.}
      \item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
      \item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaskingInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MaskingInfo-clone}{}}}
\subsection{\code{MaskingInfo$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MaskingInfo$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
