% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{MicrohaplotypeForTarget}
\alias{MicrohaplotypeForTarget}
\title{MicrohaplotypeForTarget}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
MicrohaplotypeForTarget

MicrohaplotypeForTarget
}
\details{
Microhaplotype detected for a specific target.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{mhap_id}: The index for a microhaplotype for a target in the representative_microhaplotypes list, e.g. representative_microhaplotypes[mhaps_target_id][mhap_id].
\item \code{reads}: The read count for this microhaplotype.
\item \code{umis}: The unique molecular identifier (umi) count for this microhaplotype.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mhap_id}}{The index for a microhaplotype for a target in the representative_microhaplotypes list, e.g. representative_microhaplotypes[mhaps_target_id][mhap_id].}

\item{\code{reads}}{The read count for this microhaplotype.}

\item{\code{umis}}{The unique molecular identifier (umi) count for this microhaplotype.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MicrohaplotypeForTarget-new}{\code{MicrohaplotypeForTarget$new()}}
\item \href{#method-MicrohaplotypeForTarget-validate}{\code{MicrohaplotypeForTarget$validate()}}
\item \href{#method-MicrohaplotypeForTarget-to_list}{\code{MicrohaplotypeForTarget$to_list()}}
\item \href{#method-MicrohaplotypeForTarget-to_json_list}{\code{MicrohaplotypeForTarget$to_json_list()}}
\item \href{#method-MicrohaplotypeForTarget-to_json}{\code{MicrohaplotypeForTarget$to_json()}}
\item \href{#method-MicrohaplotypeForTarget-clone}{\code{MicrohaplotypeForTarget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicrohaplotypeForTarget-new"></a>}}
\if{latex}{\out{\hypertarget{method-MicrohaplotypeForTarget-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicrohaplotypeForTarget$new(
  mhap_id = NA_real_,
  reads = NA_real_,
  umis = NULL,
  extras = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mhap_id}}{The index for a microhaplotype for a target in the representative_microhaplotypes list, e.g. representative_microhaplotypes[mhaps_target_id][mhap_id].}

\item{\code{reads}}{The read count for this microhaplotype.}

\item{\code{umis}}{The unique molecular identifier (umi) count for this microhaplotype.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicrohaplotypeForTarget-validate"></a>}}
\if{latex}{\out{\hypertarget{method-MicrohaplotypeForTarget-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicrohaplotypeForTarget$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicrohaplotypeForTarget-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-MicrohaplotypeForTarget-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicrohaplotypeForTarget$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicrohaplotypeForTarget-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-MicrohaplotypeForTarget-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicrohaplotypeForTarget$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicrohaplotypeForTarget-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-MicrohaplotypeForTarget-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicrohaplotypeForTarget$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MicrohaplotypeForTarget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MicrohaplotypeForTarget-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MicrohaplotypeForTarget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
