% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{PlateInfo}
\alias{PlateInfo}
\title{PlateInfo}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
PlateInfo

PlateInfo
}
\details{
Information about a plate location, e.g. a standard 96 well plate with row having a letter and column having a number.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{plate_col}: The column position.
\item \code{plate_name}: A name for the plate.
\item \code{plate_row}: The row position.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{plate_col}}{The column position.}

\item{\code{plate_name}}{A name for the plate.}

\item{\code{plate_row}}{The row position.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlateInfo-new}{\code{PlateInfo$new()}}
\item \href{#method-PlateInfo-validate}{\code{PlateInfo$validate()}}
\item \href{#method-PlateInfo-to_list}{\code{PlateInfo$to_list()}}
\item \href{#method-PlateInfo-to_json_list}{\code{PlateInfo$to_json_list()}}
\item \href{#method-PlateInfo-to_json}{\code{PlateInfo$to_json()}}
\item \href{#method-PlateInfo-clone}{\code{PlateInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-PlateInfo-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateInfo$new(
  plate_col = NA_real_,
  plate_name = NA_character_,
  plate_row = NA_character_,
  extras = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plate_col}}{The column position.}

\item{\code{plate_name}}{A name for the plate.}

\item{\code{plate_row}}{The row position.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateInfo-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PlateInfo-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateInfo$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateInfo-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-PlateInfo-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateInfo$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateInfo-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-PlateInfo-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateInfo$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateInfo-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-PlateInfo-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateInfo$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlateInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlateInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlateInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
