% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{PortableMicrohaplotypeObject}
\alias{PortableMicrohaplotypeObject}
\title{PortableMicrohaplotypeObject}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
PortableMicrohaplotypeObject

PortableMicrohaplotypeObject
}
\details{
Information on final microhaplotype results from a targeted amplicon analysis with associated meta data.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{bioinformatics_methods_info}: The bioinformatics pipeline/methods used to generated the microhaplotype analysis for this project.
\item \code{bioinformatics_run_info}: The runtime info for the bioinformatics pipeline used to generated the microhaplotypes analysis for this project.
\item \code{detected_microhaplotypes}: The microhaplotypes detected in this projects.
\item \code{library_sample_info}: A list of libraries of all the seq/amp of the specimens within this PMO file.
\item \code{panel_info}: A list of info on the panels.
\item \code{pmo_header}: The PMO information for this file including version etc.
\item \code{project_info}: The information about the projects stored in this PMO.
\item \code{read_counts_by_stage}: The read counts for library_samples for different stages of the pipeline.
\item \code{representative_microhaplotypes}: A list of the information on the representative microhaplotypes.
\item \code{sequencing_info}: A list of sequencing infos for this PMO file.
\item \code{specimen_info}: A list of all the specimens within this PMO file.
\item \code{target_info}: A list of info on the targets.
\item \code{targeted_genomes}: A list of genomes that any genomic location information refers to.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bioinformatics_methods_info}}{The bioinformatics pipeline/methods used to generated the microhaplotype analysis for this project.}

\item{\code{bioinformatics_run_info}}{The runtime info for the bioinformatics pipeline used to generated the microhaplotypes analysis for this project.}

\item{\code{detected_microhaplotypes}}{The microhaplotypes detected in this projects.}

\item{\code{library_sample_info}}{A list of libraries of all the seq/amp of the specimens within this PMO file.}

\item{\code{panel_info}}{A list of info on the panels.}

\item{\code{pmo_header}}{The PMO information for this file including version etc.}

\item{\code{project_info}}{The information about the projects stored in this PMO.}

\item{\code{read_counts_by_stage}}{The read counts for library_samples for different stages of the pipeline.}

\item{\code{representative_microhaplotypes}}{A list of the information on the representative microhaplotypes.}

\item{\code{sequencing_info}}{A list of sequencing infos for this PMO file.}

\item{\code{specimen_info}}{A list of all the specimens within this PMO file.}

\item{\code{target_info}}{A list of info on the targets.}

\item{\code{targeted_genomes}}{A list of genomes that any genomic location information refers to.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PortableMicrohaplotypeObject-new}{\code{PortableMicrohaplotypeObject$new()}}
\item \href{#method-PortableMicrohaplotypeObject-validate}{\code{PortableMicrohaplotypeObject$validate()}}
\item \href{#method-PortableMicrohaplotypeObject-to_list}{\code{PortableMicrohaplotypeObject$to_list()}}
\item \href{#method-PortableMicrohaplotypeObject-to_json_list}{\code{PortableMicrohaplotypeObject$to_json_list()}}
\item \href{#method-PortableMicrohaplotypeObject-to_json}{\code{PortableMicrohaplotypeObject$to_json()}}
\item \href{#method-PortableMicrohaplotypeObject-to_file}{\code{PortableMicrohaplotypeObject$to_file()}}
\item \href{#method-PortableMicrohaplotypeObject-clone}{\code{PortableMicrohaplotypeObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PortableMicrohaplotypeObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-PortableMicrohaplotypeObject-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PortableMicrohaplotypeObject$new(
  bioinformatics_methods_info = NULL,
  bioinformatics_run_info = NULL,
  detected_microhaplotypes = list(),
  library_sample_info = list(),
  panel_info = list(),
  pmo_header = NULL,
  project_info = NULL,
  read_counts_by_stage = NULL,
  representative_microhaplotypes = NULL,
  sequencing_info = NULL,
  specimen_info = list(),
  target_info = list(),
  targeted_genomes = NULL,
  extras = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bioinformatics_methods_info}}{The bioinformatics pipeline/methods used to generated the microhaplotype analysis for this project.}

\item{\code{bioinformatics_run_info}}{The runtime info for the bioinformatics pipeline used to generated the microhaplotypes analysis for this project.}

\item{\code{detected_microhaplotypes}}{The microhaplotypes detected in this projects.}

\item{\code{library_sample_info}}{A list of libraries of all the seq/amp of the specimens within this PMO file.}

\item{\code{panel_info}}{A list of info on the panels.}

\item{\code{pmo_header}}{The PMO information for this file including version etc.}

\item{\code{project_info}}{The information about the projects stored in this PMO.}

\item{\code{read_counts_by_stage}}{The read counts for library_samples for different stages of the pipeline.}

\item{\code{representative_microhaplotypes}}{A list of the information on the representative microhaplotypes.}

\item{\code{sequencing_info}}{A list of sequencing infos for this PMO file.}

\item{\code{specimen_info}}{A list of all the specimens within this PMO file.}

\item{\code{target_info}}{A list of info on the targets.}

\item{\code{targeted_genomes}}{A list of genomes that any genomic location information refers to.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PortableMicrohaplotypeObject-validate"></a>}}
\if{latex}{\out{\hypertarget{method-PortableMicrohaplotypeObject-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PortableMicrohaplotypeObject$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PortableMicrohaplotypeObject-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-PortableMicrohaplotypeObject-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PortableMicrohaplotypeObject$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PortableMicrohaplotypeObject-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-PortableMicrohaplotypeObject-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PortableMicrohaplotypeObject$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PortableMicrohaplotypeObject-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-PortableMicrohaplotypeObject-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PortableMicrohaplotypeObject$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PortableMicrohaplotypeObject-to_file"></a>}}
\if{latex}{\out{\hypertarget{method-PortableMicrohaplotypeObject-to_file}{}}}
\subsection{Method \code{to_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PortableMicrohaplotypeObject$to_file(
  path,
  pretty = FALSE,
  auto_unbox = TRUE,
  validate = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PortableMicrohaplotypeObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PortableMicrohaplotypeObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PortableMicrohaplotypeObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
