% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{Pseudocigar}
\alias{Pseudocigar}
\title{Pseudocigar}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
Pseudocigar

Pseudocigar
}
\details{
Information on pseudocigar for a sequence.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{pseudocigar_generation_description}: A description of how the pseudocigar information was generated.
\item \code{pseudocigar_seq}: The pseudocigar itself.
\item \code{ref_loc}: The genomic location the pseudocigar is in reference to.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pseudocigar_generation_description}}{A description of how the pseudocigar information was generated.}

\item{\code{pseudocigar_seq}}{The pseudocigar itself.}

\item{\code{ref_loc}}{The genomic location the pseudocigar is in reference to.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Pseudocigar-new}{\code{Pseudocigar$new()}}
\item \href{#method-Pseudocigar-validate}{\code{Pseudocigar$validate()}}
\item \href{#method-Pseudocigar-to_list}{\code{Pseudocigar$to_list()}}
\item \href{#method-Pseudocigar-to_json_list}{\code{Pseudocigar$to_json_list()}}
\item \href{#method-Pseudocigar-to_json}{\code{Pseudocigar$to_json()}}
\item \href{#method-Pseudocigar-clone}{\code{Pseudocigar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pseudocigar-new"></a>}}
\if{latex}{\out{\hypertarget{method-Pseudocigar-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pseudocigar$new(
  pseudocigar_generation_description = NULL,
  pseudocigar_seq = NA_character_,
  ref_loc = NULL,
  extras = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pseudocigar_generation_description}}{A description of how the pseudocigar information was generated.}

\item{\code{pseudocigar_seq}}{The pseudocigar itself.}

\item{\code{ref_loc}}{The genomic location the pseudocigar is in reference to.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pseudocigar-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Pseudocigar-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pseudocigar$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pseudocigar-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Pseudocigar-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pseudocigar$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pseudocigar-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-Pseudocigar-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pseudocigar$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pseudocigar-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-Pseudocigar-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pseudocigar$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Pseudocigar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Pseudocigar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Pseudocigar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
