% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{ReadCountsByStage}
\alias{ReadCountsByStage}
\title{ReadCountsByStage}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
ReadCountsByStage

ReadCountsByStage
}
\details{
Information on the reads counts at several stages of a pipeline.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{bioinformatics_run_id}: The index into bioinformatics_run_info list.
\item \code{read_counts_by_library_sample_by_stage}: A list by library_sample for the counts at each stage.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{bioinformatics_run_id}}{The index into bioinformatics_run_info list.}

\item{\code{read_counts_by_library_sample_by_stage}}{A list by library_sample for the counts at each stage.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReadCountsByStage-new}{\code{ReadCountsByStage$new()}}
\item \href{#method-ReadCountsByStage-validate}{\code{ReadCountsByStage$validate()}}
\item \href{#method-ReadCountsByStage-to_list}{\code{ReadCountsByStage$to_list()}}
\item \href{#method-ReadCountsByStage-to_json_list}{\code{ReadCountsByStage$to_json_list()}}
\item \href{#method-ReadCountsByStage-to_json}{\code{ReadCountsByStage$to_json()}}
\item \href{#method-ReadCountsByStage-clone}{\code{ReadCountsByStage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadCountsByStage-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReadCountsByStage-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadCountsByStage$new(
  bioinformatics_run_id = NULL,
  read_counts_by_library_sample_by_stage = list(),
  extras = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bioinformatics_run_id}}{The index into bioinformatics_run_info list.}

\item{\code{read_counts_by_library_sample_by_stage}}{A list by library_sample for the counts at each stage.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadCountsByStage-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ReadCountsByStage-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadCountsByStage$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadCountsByStage-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ReadCountsByStage-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadCountsByStage$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadCountsByStage-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-ReadCountsByStage-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadCountsByStage$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadCountsByStage-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-ReadCountsByStage-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadCountsByStage$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadCountsByStage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReadCountsByStage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadCountsByStage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
