% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{RepresentativeMicrohaplotype}
\alias{RepresentativeMicrohaplotype}
\title{RepresentativeMicrohaplotype}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
RepresentativeMicrohaplotype

RepresentativeMicrohaplotype
}
\details{
The representative sequence for a microhaplotype.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{alt_annotations}: A list of additional annotations associated with this microhaplotype, e.g. wildtype.
\item \code{associated_protein_variants}: A list of protein variants for this haplotype, e.g. amino acid changes/INDELS.
\item \code{associated_seq_variants}: A list of sequence variants for this haplotype, e.g. SNPS, indels.
\item \code{masking}: Masking info for the sequence.
\item \code{microhaplotype_name}: An optional name for this microhaplotype.
\item \code{pseudocigar}: The pseudocigar of the haplotype.
\item \code{quality}: The ASCII fastq per base quality score for this sequence, this is optional, must be same length as the sequence.
\item \code{seq}: The sequence.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alt_annotations}}{A list of additional annotations associated with this microhaplotype, e.g. wildtype.}

\item{\code{associated_protein_variants}}{A list of protein variants for this haplotype, e.g. amino acid changes/INDELS.}

\item{\code{associated_seq_variants}}{A list of sequence variants for this haplotype, e.g. SNPS, indels.}

\item{\code{masking}}{Masking info for the sequence.}

\item{\code{microhaplotype_name}}{An optional name for this microhaplotype.}

\item{\code{pseudocigar}}{The pseudocigar of the haplotype.}

\item{\code{quality}}{The ASCII fastq per base quality score for this sequence, this is optional, must be same length as the sequence.}

\item{\code{seq}}{The sequence.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RepresentativeMicrohaplotype-new}{\code{RepresentativeMicrohaplotype$new()}}
\item \href{#method-RepresentativeMicrohaplotype-validate}{\code{RepresentativeMicrohaplotype$validate()}}
\item \href{#method-RepresentativeMicrohaplotype-to_list}{\code{RepresentativeMicrohaplotype$to_list()}}
\item \href{#method-RepresentativeMicrohaplotype-to_json_list}{\code{RepresentativeMicrohaplotype$to_json_list()}}
\item \href{#method-RepresentativeMicrohaplotype-to_json}{\code{RepresentativeMicrohaplotype$to_json()}}
\item \href{#method-RepresentativeMicrohaplotype-clone}{\code{RepresentativeMicrohaplotype$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentativeMicrohaplotype-new"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentativeMicrohaplotype-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentativeMicrohaplotype$new(
  alt_annotations = NULL,
  associated_protein_variants = NULL,
  associated_seq_variants = NULL,
  masking = NULL,
  microhaplotype_name = NULL,
  pseudocigar = NULL,
  quality = NULL,
  seq = NA_character_,
  extras = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alt_annotations}}{A list of additional annotations associated with this microhaplotype, e.g. wildtype.}

\item{\code{associated_protein_variants}}{A list of protein variants for this haplotype, e.g. amino acid changes/INDELS.}

\item{\code{associated_seq_variants}}{A list of sequence variants for this haplotype, e.g. SNPS, indels.}

\item{\code{masking}}{Masking info for the sequence.}

\item{\code{microhaplotype_name}}{An optional name for this microhaplotype.}

\item{\code{pseudocigar}}{The pseudocigar of the haplotype.}

\item{\code{quality}}{The ASCII fastq per base quality score for this sequence, this is optional, must be same length as the sequence.}

\item{\code{seq}}{The sequence.}

\item{\code{extras}}{Additional properties not explicitly defined in the schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentativeMicrohaplotype-validate"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentativeMicrohaplotype-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the current instance against schema-derived constraints.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentativeMicrohaplotype$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentativeMicrohaplotype-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentativeMicrohaplotype-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert the object to a plain R list using in-memory values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentativeMicrohaplotype$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentativeMicrohaplotype-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentativeMicrohaplotype-to_json_list}{}}}
\subsection{Method \code{to_json_list()}}{
Convert the object to a JSON-ready R list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentativeMicrohaplotype$to_json_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentativeMicrohaplotype-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentativeMicrohaplotype-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert the object to a JSON string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentativeMicrohaplotype$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Logical; pretty-print the JSON.}

\item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}

\item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RepresentativeMicrohaplotype-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RepresentativeMicrohaplotype-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RepresentativeMicrohaplotype$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
