% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{SequencingInfo}
\alias{SequencingInfo}
\title{SequencingInfo}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
Information on sequencing info.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{library_kit}: Name, version, and applicable cell or cycle numbers for the kit used to prepare libraries and load cells or chips for sequencing. If possible, include a part number, e.g. MiSeq Reagent Kit v3 (150-cycle), MS-102-3001.
\item \code{library_layout}: Specify the configuration of reads, e.g. paired-end, single.
\item \code{library_screen}: Describe enrichment, screening, or normalization methods applied during amplification or library preparation, e.g. size selection 390bp, diluted to 1 ng DNA/sample.
\item \code{library_selection}: How amplification was done (common are PCR=Source material was selected by designed primers, RANDOM =Random selection by shearing or other method).
\item \code{library_source}: Source of amplification material e.g. was it DNA (GENOMIC) or RNA (TRANSCRIPTOMIC) (common names GENOMIC, TRANSCRIPTOMIC).
\item \code{library_strategy}: What the nuceloacid sequencing/amplification strategy was (common names are AMPLICON, WGS).
\item \code{nucl_acid_amp}: Link to a reference or kit that describes the enzymatic amplification of nucleic acids.
\item \code{nucl_acid_amp_date}: The date of the nucleoacid amplification.
\item \code{nucl_acid_ext}: Link to a reference or kit that describes the recovery of nucleic acids from the sample.
\item \code{nucl_acid_ext_date}: The date of the nucleoacid extraction.
\item \code{pcr_cond}: The method/conditions for PCR, List PCR cycles used to amplify the target.
\item \code{seq_center}: Name of facility where sequencing was performed (lab, core facility, or company).
\item \code{seq_date}: The date of sequencing, should be YYYY-MM or YYYY-MM-DD.
\item \code{seq_instrument_model}: The sequencing instrument model used to sequence the run, e.g. NextSeq 2000, MinION, Revio.
\item \code{seq_platform}: The sequencing technology used to sequence the run, e.g. ILLUMINA, NANOPORE, PACBIO.
\item \code{sequencing_info_name}: A name for a specific sequencing run, e.g. batch1.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{library_kit}}{Name, version, and applicable cell or cycle numbers for the kit used to prepare libraries and load cells or chips for sequencing. If possible, include a part number, e.g. MiSeq Reagent Kit v3 (150-cycle), MS-102-3001.}

    \item{\code{library_layout}}{Specify the configuration of reads, e.g. paired-end, single.}

    \item{\code{library_screen}}{Describe enrichment, screening, or normalization methods applied during amplification or library preparation, e.g. size selection 390bp, diluted to 1 ng DNA/sample.}

    \item{\code{library_selection}}{How amplification was done (common are PCR=Source material was selected by designed primers, RANDOM =Random selection by shearing or other method).}

    \item{\code{library_source}}{Source of amplification material e.g. was it DNA (GENOMIC) or RNA (TRANSCRIPTOMIC) (common names GENOMIC, TRANSCRIPTOMIC).}

    \item{\code{library_strategy}}{What the nuceloacid sequencing/amplification strategy was (common names are AMPLICON, WGS).}

    \item{\code{nucl_acid_amp}}{Link to a reference or kit that describes the enzymatic amplification of nucleic acids.}

    \item{\code{nucl_acid_amp_date}}{The date of the nucleoacid amplification.}

    \item{\code{nucl_acid_ext}}{Link to a reference or kit that describes the recovery of nucleic acids from the sample.}

    \item{\code{nucl_acid_ext_date}}{The date of the nucleoacid extraction.}

    \item{\code{pcr_cond}}{The method/conditions for PCR, List PCR cycles used to amplify the target.}

    \item{\code{seq_center}}{Name of facility where sequencing was performed (lab, core facility, or company).}

    \item{\code{seq_date}}{The date of sequencing, should be YYYY-MM or YYYY-MM-DD.}

    \item{\code{seq_instrument_model}}{The sequencing instrument model used to sequence the run, e.g. NextSeq 2000, MinION, Revio.}

    \item{\code{seq_platform}}{The sequencing technology used to sequence the run, e.g. ILLUMINA, NANOPORE, PACBIO.}

    \item{\code{sequencing_info_name}}{A name for a specific sequencing run, e.g. batch1.}

    \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-SequencingInfo-initialize}{\code{SequencingInfo$new()}}
    \item \href{#method-SequencingInfo-validate}{\code{SequencingInfo$validate()}}
    \item \href{#method-SequencingInfo-to_list}{\code{SequencingInfo$to_list()}}
    \item \href{#method-SequencingInfo-to_json_list}{\code{SequencingInfo$to_json_list()}}
    \item \href{#method-SequencingInfo-to_json}{\code{SequencingInfo$to_json()}}
    \item \href{#method-SequencingInfo-clone}{\code{SequencingInfo$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SequencingInfo-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-SequencingInfo-initialize}{}}}
\subsection{\code{SequencingInfo$new()}}{
  Create a new instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SequencingInfo$new(
  library_kit = NULL,
  library_layout = NA_character_,
  library_screen = NULL,
  library_selection = NA_character_,
  library_source = NA_character_,
  library_strategy = NA_character_,
  nucl_acid_amp = NULL,
  nucl_acid_amp_date = NULL,
  nucl_acid_ext = NULL,
  nucl_acid_ext_date = NULL,
  pcr_cond = NULL,
  seq_center = NULL,
  seq_date = NULL,
  seq_instrument_model = NA_character_,
  seq_platform = NA_character_,
  sequencing_info_name = NA_character_,
  extras = list()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{library_kit}}{Name, version, and applicable cell or cycle numbers for the kit used to prepare libraries and load cells or chips for sequencing. If possible, include a part number, e.g. MiSeq Reagent Kit v3 (150-cycle), MS-102-3001.}
      \item{\code{library_layout}}{Specify the configuration of reads, e.g. paired-end, single.}
      \item{\code{library_screen}}{Describe enrichment, screening, or normalization methods applied during amplification or library preparation, e.g. size selection 390bp, diluted to 1 ng DNA/sample.}
      \item{\code{library_selection}}{How amplification was done (common are PCR=Source material was selected by designed primers, RANDOM =Random selection by shearing or other method).}
      \item{\code{library_source}}{Source of amplification material e.g. was it DNA (GENOMIC) or RNA (TRANSCRIPTOMIC) (common names GENOMIC, TRANSCRIPTOMIC).}
      \item{\code{library_strategy}}{What the nuceloacid sequencing/amplification strategy was (common names are AMPLICON, WGS).}
      \item{\code{nucl_acid_amp}}{Link to a reference or kit that describes the enzymatic amplification of nucleic acids.}
      \item{\code{nucl_acid_amp_date}}{The date of the nucleoacid amplification.}
      \item{\code{nucl_acid_ext}}{Link to a reference or kit that describes the recovery of nucleic acids from the sample.}
      \item{\code{nucl_acid_ext_date}}{The date of the nucleoacid extraction.}
      \item{\code{pcr_cond}}{The method/conditions for PCR, List PCR cycles used to amplify the target.}
      \item{\code{seq_center}}{Name of facility where sequencing was performed (lab, core facility, or company).}
      \item{\code{seq_date}}{The date of sequencing, should be YYYY-MM or YYYY-MM-DD.}
      \item{\code{seq_instrument_model}}{The sequencing instrument model used to sequence the run, e.g. NextSeq 2000, MinION, Revio.}
      \item{\code{seq_platform}}{The sequencing technology used to sequence the run, e.g. ILLUMINA, NANOPORE, PACBIO.}
      \item{\code{sequencing_info_name}}{A name for a specific sequencing run, e.g. batch1.}
      \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SequencingInfo-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SequencingInfo-validate}{}}}
\subsection{\code{SequencingInfo$validate()}}{
  Validate the current instance against schema-derived constraints.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SequencingInfo$validate()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SequencingInfo-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SequencingInfo-to_list}{}}}
\subsection{\code{SequencingInfo$to_list()}}{
  Convert the object to a plain R list using in-memory values.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SequencingInfo$to_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SequencingInfo-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-SequencingInfo-to_json_list}{}}}
\subsection{\code{SequencingInfo$to_json_list()}}{
  Convert the object to a JSON-ready R list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SequencingInfo$to_json_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SequencingInfo-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-SequencingInfo-to_json}{}}}
\subsection{\code{SequencingInfo$to_json()}}{
  Convert the object to a JSON string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SequencingInfo$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pretty}}{Logical; pretty-print the JSON.}
      \item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
      \item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SequencingInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SequencingInfo-clone}{}}}
\subsection{\code{SequencingInfo$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SequencingInfo$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
