% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmo_r6.R
\name{SpecimenInfo}
\alias{SpecimenInfo}
\title{SpecimenInfo}
\format{
An \code{\link[R6:R6Class]{R6::R6Class()}} generator object.
}
\description{
Information on specimen info.

Auto-generated R6 class from JSON Schema.
}
\section{Constructor}{

\code{new(...)} supports the following arguments.
\itemize{
\item \code{alternate_identifiers}: A list of alternative names.
\item \code{blood_meal}: Whether host specimen has had a recent blood meal.
\item \code{collection_country}: The name of country collected in, would be the same as admin level 0.
\item \code{collection_date}: The date of the specimen collection, can be YYYY, YYYY-MM, or YYYY-MM-DD.
\item \code{drug_usage}: Any drug used by subject and the frequency of usage; can include multiple drugs used.
\item \code{env_broad_scale}: The broad environment from which the specimen was collected, e.g. highlands, lowlands, mountainous region.
\item \code{env_local_scale}: The local environment from which the specimen was collected, e.g. jungle, urban, rural.
\item \code{env_medium}: The environment medium from which the specimen was collected from.
\item \code{geo_admin1}: Geographical admin level 1, the secondary large demarcation of a nation (nation = admin level 0).
\item \code{geo_admin2}: Geographical admin level 2, the third large demarcation of a nation (nation = admin level 0).
\item \code{geo_admin3}: Geographical admin level 3, the third large demarcation of a nation (nation = admin level 0).
\item \code{gravid}: Whether host specimen is currently pregnant.
\item \code{gravidity}: The gravidity of the specimen host (number of previous pregnancies).
\item \code{has_travel_out_six_month}: Has travelled out from local region in the last six months.
\item \code{host_age}: If specimen is from a person, the age in years of the person, can be float value so for 3 month old put 0.25.
\item \code{host_sex}: If specimen is collected from a host with a sex, the sex listed for that host.
\item \code{host_subject_name}: An identifier for the individual/person/patient a specimen was collected from.
\item \code{host_taxon_id}: The NCBI taxonomy number of the host that the specimen was collected from.
\item \code{lat_lon}: The latitude and longitude of a specific site.
\item \code{parasite_density_info}: One or more parasite densities in microliters for this specimen.
\item \code{project_id}: The index into the project_info list.
\item \code{specimen_accession}: If specimen is deposited in a database, what accession is it associated with.
\item \code{specimen_collect_device}: The way the specimen was collected, e.g. whole blood, dried blood spot.
\item \code{specimen_comments}: Any additional comments about the specimen.
\item \code{specimen_name}: An identifier for the specimen, should be unique within this sample set.
\item \code{specimen_store_loc}: The specimen store site, address or facility name.
\item \code{specimen_taxon_id}: The NCBI taxonomy number of the organism(s) in the specimen, can list multiple if a mixed sample.
\item \code{specimen_type}: What type of specimen this is, e.g. negative_control, positive_control, field_sample.
\item \code{storage_plate_info}: Plate location of where specimen is stored if stored in a plate.
\item \code{travel_out_six_month}: Specification of the countries travelled in the last six months; can include multiple travels.
\item \code{treatment_status}: If person has been treated with drugs, what was the treatment outcome.
\item \code{extras}: Additional properties not explicitly defined in the schema.
}
}

\section{Methods}{

See inline method documentation for \code{initialize()}, \code{validate()}, \code{to_list()}, \code{to_json_list()}, and \code{to_json()}.
}

\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{alternate_identifiers}}{A list of alternative names.}

    \item{\code{blood_meal}}{Whether host specimen has had a recent blood meal.}

    \item{\code{collection_country}}{The name of country collected in, would be the same as admin level 0.}

    \item{\code{collection_date}}{The date of the specimen collection, can be YYYY, YYYY-MM, or YYYY-MM-DD.}

    \item{\code{drug_usage}}{Any drug used by subject and the frequency of usage; can include multiple drugs used.}

    \item{\code{env_broad_scale}}{The broad environment from which the specimen was collected, e.g. highlands, lowlands, mountainous region.}

    \item{\code{env_local_scale}}{The local environment from which the specimen was collected, e.g. jungle, urban, rural.}

    \item{\code{env_medium}}{The environment medium from which the specimen was collected from.}

    \item{\code{geo_admin1}}{Geographical admin level 1, the secondary large demarcation of a nation (nation = admin level 0).}

    \item{\code{geo_admin2}}{Geographical admin level 2, the third large demarcation of a nation (nation = admin level 0).}

    \item{\code{geo_admin3}}{Geographical admin level 3, the third large demarcation of a nation (nation = admin level 0).}

    \item{\code{gravid}}{Whether host specimen is currently pregnant.}

    \item{\code{gravidity}}{The gravidity of the specimen host (number of previous pregnancies).}

    \item{\code{has_travel_out_six_month}}{Has travelled out from local region in the last six months.}

    \item{\code{host_age}}{If specimen is from a person, the age in years of the person, can be float value so for 3 month old put 0.25.}

    \item{\code{host_sex}}{If specimen is collected from a host with a sex, the sex listed for that host.}

    \item{\code{host_subject_name}}{An identifier for the individual/person/patient a specimen was collected from.}

    \item{\code{host_taxon_id}}{The NCBI taxonomy number of the host that the specimen was collected from.}

    \item{\code{lat_lon}}{The latitude and longitude of a specific site.}

    \item{\code{parasite_density_info}}{One or more parasite densities in microliters for this specimen.}

    \item{\code{project_id}}{The index into the project_info list.}

    \item{\code{specimen_accession}}{If specimen is deposited in a database, what accession is it associated with.}

    \item{\code{specimen_collect_device}}{The way the specimen was collected, e.g. whole blood, dried blood spot.}

    \item{\code{specimen_comments}}{Any additional comments about the specimen.}

    \item{\code{specimen_name}}{An identifier for the specimen, should be unique within this sample set.}

    \item{\code{specimen_store_loc}}{The specimen store site, address or facility name.}

    \item{\code{specimen_taxon_id}}{The NCBI taxonomy number of the organism(s) in the specimen, can list multiple if a mixed sample.}

    \item{\code{specimen_type}}{What type of specimen this is, e.g. negative_control, positive_control, field_sample.}

    \item{\code{storage_plate_info}}{Plate location of where specimen is stored if stored in a plate.}

    \item{\code{travel_out_six_month}}{Specification of the countries travelled in the last six months; can include multiple travels.}

    \item{\code{treatment_status}}{If person has been treated with drugs, what was the treatment outcome.}

    \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-SpecimenInfo-initialize}{\code{SpecimenInfo$new()}}
    \item \href{#method-SpecimenInfo-validate}{\code{SpecimenInfo$validate()}}
    \item \href{#method-SpecimenInfo-to_list}{\code{SpecimenInfo$to_list()}}
    \item \href{#method-SpecimenInfo-to_json_list}{\code{SpecimenInfo$to_json_list()}}
    \item \href{#method-SpecimenInfo-to_json}{\code{SpecimenInfo$to_json()}}
    \item \href{#method-SpecimenInfo-clone}{\code{SpecimenInfo$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenInfo-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenInfo-initialize}{}}}
\subsection{\code{SpecimenInfo$new()}}{
  Create a new instance.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenInfo$new(
  alternate_identifiers = NULL,
  blood_meal = NULL,
  collection_country = NULL,
  collection_date = NULL,
  drug_usage = NULL,
  env_broad_scale = NULL,
  env_local_scale = NULL,
  env_medium = NULL,
  geo_admin1 = NULL,
  geo_admin2 = NULL,
  geo_admin3 = NULL,
  gravid = NULL,
  gravidity = NULL,
  has_travel_out_six_month = NULL,
  host_age = NULL,
  host_sex = NULL,
  host_subject_name = NULL,
  host_taxon_id = NULL,
  lat_lon = NULL,
  parasite_density_info = NULL,
  project_id = NULL,
  specimen_accession = NULL,
  specimen_collect_device = NULL,
  specimen_comments = NULL,
  specimen_name = NA_character_,
  specimen_store_loc = NULL,
  specimen_taxon_id = NULL,
  specimen_type = NULL,
  storage_plate_info = NULL,
  travel_out_six_month = NULL,
  treatment_status = NULL,
  extras = list()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{alternate_identifiers}}{A list of alternative names.}
      \item{\code{blood_meal}}{Whether host specimen has had a recent blood meal.}
      \item{\code{collection_country}}{The name of country collected in, would be the same as admin level 0.}
      \item{\code{collection_date}}{The date of the specimen collection, can be YYYY, YYYY-MM, or YYYY-MM-DD.}
      \item{\code{drug_usage}}{Any drug used by subject and the frequency of usage; can include multiple drugs used.}
      \item{\code{env_broad_scale}}{The broad environment from which the specimen was collected, e.g. highlands, lowlands, mountainous region.}
      \item{\code{env_local_scale}}{The local environment from which the specimen was collected, e.g. jungle, urban, rural.}
      \item{\code{env_medium}}{The environment medium from which the specimen was collected from.}
      \item{\code{geo_admin1}}{Geographical admin level 1, the secondary large demarcation of a nation (nation = admin level 0).}
      \item{\code{geo_admin2}}{Geographical admin level 2, the third large demarcation of a nation (nation = admin level 0).}
      \item{\code{geo_admin3}}{Geographical admin level 3, the third large demarcation of a nation (nation = admin level 0).}
      \item{\code{gravid}}{Whether host specimen is currently pregnant.}
      \item{\code{gravidity}}{The gravidity of the specimen host (number of previous pregnancies).}
      \item{\code{has_travel_out_six_month}}{Has travelled out from local region in the last six months.}
      \item{\code{host_age}}{If specimen is from a person, the age in years of the person, can be float value so for 3 month old put 0.25.}
      \item{\code{host_sex}}{If specimen is collected from a host with a sex, the sex listed for that host.}
      \item{\code{host_subject_name}}{An identifier for the individual/person/patient a specimen was collected from.}
      \item{\code{host_taxon_id}}{The NCBI taxonomy number of the host that the specimen was collected from.}
      \item{\code{lat_lon}}{The latitude and longitude of a specific site.}
      \item{\code{parasite_density_info}}{One or more parasite densities in microliters for this specimen.}
      \item{\code{project_id}}{The index into the project_info list.}
      \item{\code{specimen_accession}}{If specimen is deposited in a database, what accession is it associated with.}
      \item{\code{specimen_collect_device}}{The way the specimen was collected, e.g. whole blood, dried blood spot.}
      \item{\code{specimen_comments}}{Any additional comments about the specimen.}
      \item{\code{specimen_name}}{An identifier for the specimen, should be unique within this sample set.}
      \item{\code{specimen_store_loc}}{The specimen store site, address or facility name.}
      \item{\code{specimen_taxon_id}}{The NCBI taxonomy number of the organism(s) in the specimen, can list multiple if a mixed sample.}
      \item{\code{specimen_type}}{What type of specimen this is, e.g. negative_control, positive_control, field_sample.}
      \item{\code{storage_plate_info}}{Plate location of where specimen is stored if stored in a plate.}
      \item{\code{travel_out_six_month}}{Specification of the countries travelled in the last six months; can include multiple travels.}
      \item{\code{treatment_status}}{If person has been treated with drugs, what was the treatment outcome.}
      \item{\code{extras}}{Additional properties not explicitly defined in the schema.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenInfo-validate"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenInfo-validate}{}}}
\subsection{\code{SpecimenInfo$validate()}}{
  Validate the current instance against schema-derived constraints.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenInfo$validate()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenInfo-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenInfo-to_list}{}}}
\subsection{\code{SpecimenInfo$to_list()}}{
  Convert the object to a plain R list using in-memory values.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenInfo$to_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenInfo-to_json_list"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenInfo-to_json_list}{}}}
\subsection{\code{SpecimenInfo$to_json_list()}}{
  Convert the object to a JSON-ready R list.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenInfo$to_json_list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenInfo-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenInfo-to_json}{}}}
\subsection{\code{SpecimenInfo$to_json()}}{
  Convert the object to a JSON string.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenInfo$to_json(pretty = FALSE, auto_unbox = TRUE, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{pretty}}{Logical; pretty-print the JSON.}
      \item{\code{auto_unbox}}{Logical; passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
      \item{\code{...}}{Additional arguments passed to \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpecimenInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpecimenInfo-clone}{}}}
\subsection{\code{SpecimenInfo$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{SpecimenInfo$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
